% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExample_clmdu.R
\docType{data}
\name{dataExample_clmdu}
\alias{dataExample_clmdu}
\title{Dummy data for clmdu example}
\format{
A data frame with 200 observations on the following variables:
\describe{
  \item{\code{X1}}{Continuous variable 1.}
  \item{\code{X2}}{Continuous variable 2.}
  \item{\code{X3}}{Continuous variable 3.}
  \item{\code{X4}}{Continuous variable 4.}
  \item{\code{Y1}}{Discrete variable 1.}
  \item{\code{Y2}}{Discrete variable 2.}
  \item{\code{Y3}}{Discrete variable 3.}
  \item{\code{Y4}}{Discrete variable 4.}
  \item{\code{Y5}}{Discrete variable 5.}
 }
}
\usage{
dataExample_clmdu
}
\description{
Dummy data for clmdu example
}
\keyword{datasets}
