% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExample_mru.R
\docType{data}
\name{dataExample_mru}
\alias{dataExample_mru}
\title{Dummy data for mru example}
\format{
A data frame with 234 observations on the following variables:
\describe{
  \item{\code{y}}{Categorical variable.}
  \item{\code{X1}}{Continuous variable 1.}
  \item{\code{X2}}{Continuous variable 2.}
  \item{\code{X3}}{Continuous variable 3.}
  \item{\code{X4}}{Continuous variable 4.}
  \item{\code{X5}}{Continuous variable 5.}
 }
}
\usage{
dataExample_mru
}
\description{
Dummy data for mru example
}
\keyword{datasets}
