% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bootstrap.R
\name{plot.bootstrap}
\alias{plot.bootstrap}
\title{Plot an object obtained using one of the bootstrap functions}
\arguments{
\item{x}{an object of type bootstrap}

\item{level}{level of confidence regios}

\item{type}{choose between "Bag" (default) or "norm"}

\item{\dots}{additional arguments to be passed.}

\item{prop}{Proportion of all the points to be included in the bag (default is 0.5)}
}
\value{
Plot of the results obtained from bootstrap
}
\description{
Plot an object obtained using one of the bootstrap functions
}
