% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.df.for.varlabels.R
\name{make.df.for.varlabels}
\alias{make.df.for.varlabels}
\title{Helper function for the plot functions}
\usage{
make.df.for.varlabels(BV, margins, names, P, R)
}
\arguments{
\item{BV}{Concatention of matrices B (P x S) and V (R x S) or only V (R x S) from lpca}

\item{margins}{a vector of length four indicating the margins of the plot}

\item{names}{a vector with variable names}

\item{P}{an integer indicating the number of predictor variables}

\item{R}{an integer indicating the number of response variables}
}
\value{
df with information for the placement of the variable labels
}
\description{
Helper function for the plot functions to add variable labels 
to the predictor and response variable axes
}
