% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.dfs.for.X.R
\name{make.dfs.for.X}
\alias{make.dfs.for.X}
\title{Helper function for the plot functions}
\usage{
make.dfs.for.X(Xo, P, B, xnames, mx, sdx)
}
\arguments{
\item{Xo}{Original predictor matrix}

\item{P}{an integer indicating the number of predictor variables}

\item{B}{matrix (P x S) with weights}

\item{xnames}{a vector with variable names}

\item{mx}{averages of the original predictor matrix}

\item{sdx}{standard deviations of the original predictor matrix}
}
\value{
output with information for the placement of variable markers and labels
}
\description{
Helper function for the plot functions to add variable markers and labels 
to the predictor variable axes
}
