% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesda.R
\docType{data}
\name{nesda}
\alias{nesda}
\title{Netherlands Study for Depression and Anxiety}
\format{
A list of 2 matrices
\itemize{
  \item X: A 845 x 8 matrix containing observed values on eight predictor variables.
  Gender (female = 1), Age, Education, Neuroticism, Extraversion, Openess, Agreeableness, Conscientiousness
  \item Y: A 845 x 5 matrix Indicating whether a participants suffers from a mental illness
  D (dysthymia), M (Major depressive disorder), G (Generalized Anxiety disorder), S (Social Phobia), P (Panic disorder)
}
}
\usage{
data(nesda)
}
\description{
Synthetic data generated on the basis of characteristics of the data described in 
Spinhoven e.a. (2009).
}
\references{
Penninx BWJH, Beekman ATF, Smit JH, Zitman FG, Nolen WA, Spinhoven P, et al.
The Netherlands Study of Depression and Anxiety (NESDA): rationale, objectives, and methods.
International Journal of Methods in Psychiatric Research 2008;17:121–40.

Spinhoven, P., De Rooij, M., Heiser, W.J., Penninx, B. and Smit, J. (2009).
The Role of Personality in Comorbidity among Anxiety and Depressive Disorders in Primary Care and Specialty Care: A Cross-Sectional Analysis.
General Hospital Psychiatry, 31, 470-477.
}
\keyword{dataset}
