% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.trioscale.R
\name{plot.trioscale}
\alias{plot.trioscale}
\title{Plotting function for object of class trioscale}
\usage{
\method{plot}{trioscale}(
  x,
  ycol = "darkgreen",
  xcol = "darkblue",
  ocol = "grey",
  markersize = 2.5,
  labelsize = 3,
  classlabels = NULL,
  s1 = 2.5,
  s2 = 1.05,
  s3 = 1.15,
  ...
)
}
\arguments{
\item{x}{An object of class trioscale.}

\item{ycol}{colour for representation of response variables}

\item{xcol}{colour for representation of predictor variables}

\item{ocol}{colour for representation of row objects}

\item{markersize}{size of points}

\item{labelsize}{size of labels}

\item{classlabels}{List with plotting options for the labels of the Anchor points}

\item{s1}{scaling factor for distance between points and log-ratio axes}

\item{s2}{scaling factor for positioning class labels}

\item{s3}{scaling factor for positioning variable lables}

\item{\dots}{additional arguments to be passed.}
}
\value{
This function returns an plot
}
\description{
Plotting function for object of class trioscale
}
\examples{
\dontrun{
out = trioscale(data)
plot.trioscale(out)
}


}
