% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.R
\name{procrustes1}
\alias{procrustes1}
\title{Two procedures for procrustes analysis}
\usage{
procrustes1(V1, V2)
}
\arguments{
\item{V1}{first coordinate matrix}

\item{V2}{second coordinate matrix}
}
\value{
either a rotation matrix (procrustes1) or a list with a rotation matrix (R), a stretching factor (s) and a translation (t) (procrustes2)
}
\description{
procrustes1 does orthogonal procrustes analysis
procrustes2 does similarity procrustes analysis
}
