% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_drugdata.R
\name{read_drugdata}
\alias{read_drugdata}
\title{Function for reading the drug consumption data from the UCI repository}
\usage{
read_drugdata()
}
\value{
X first coordinate matrix

Y matrix with response variables (Alcohol,Am,Amyl,Be,Caff,Ca,Choc,Co,Crack,Ex,Heroin,Ke,Le,LSD,Me,Mu,Ni,Semer,VSA)

idx indicator which response variables have a probability between 0.1 and 0.9
}
\description{
Function for reading the drug consumption data from the UCI repository
}
\references{
Fehrman, E., Muhammad, A. K., Mirkes, E. M., Egan, V., & Gorban, A. N. (2017).
The five factor model of personality and evaluation of drug consumption risk. In Data science:
innovative developments in data analysis and clustering (pp. 231-242). Springer International Publishing.
}
