% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_isspdata_peb.R
\name{read_isspdata_peb}
\alias{read_isspdata_peb}
\title{Function to read in the ISSP data
It requires the file ZA7650_v1-0-0.sav to be on your computer
this file can be obtained from
/www.gesis.org/en/issp/modules/issp-modules-by-topic/environment/2020
ZA7650 Data file Version 1.0.0, https://doi.org/10.4232/1.13921.}
\usage{
read_isspdata_peb(path)
}
\arguments{
\item{path}{path where the ZA7650_v1-0-0.sav file is saved}
}
\value{
X matrix containing the predictor variables

Y matrix with response variables
}
\description{
Function to read in the ISSP data
It requires the file ZA7650_v1-0-0.sav to be on your computer
this file can be obtained from
/www.gesis.org/en/issp/modules/issp-modules-by-topic/environment/2020
ZA7650 Data file Version 1.0.0, https://doi.org/10.4232/1.13921.
}
\references{
ISSP Research Group (2022). International social survey programme:
Environment IV - issp 2020. GESIS, Cologne.
}
