% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.clmdu.R
\name{summary.clmdu}
\alias{summary.clmdu}
\title{Summarizing Cumulative Logistic MDU models
									 
The function summary.lmdu gives a summary from an object from clmdu()}
\usage{
\method{summary}{clmdu}(object, ...)
}
\arguments{
\item{object}{An object resulting from clmdu}

\item{\dots}{additional arguments to be passed.}
}
\value{
Summary of the results obtained from clmdu
}
\description{
Summarizing Cumulative Logistic MDU models
									 
The function summary.lmdu gives a summary from an object from clmdu()
}
