% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.clpca.R
\name{summary.clpca}
\alias{summary.clpca}
\title{Summarizing Cumulative Logistic PCA models}
\usage{
\method{summary}{clpca}(object, ...)
}
\arguments{
\item{object}{An object resulting from clpca}

\item{\dots}{additional arguments to be passed.}
}
\value{
Summary of the results obtained from clpca
}
\description{
The function summary.clpca gives a summary from an object from clpca()
}
