% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.esm.R
\name{summary.esm}
\alias{summary.esm}
\title{Summarizing an Extended Steretype Model}
\usage{
\method{summary}{esm}(object, ...)
}
\arguments{
\item{object}{An object resulting from esm}

\item{\dots}{additional arguments to be passed.}
}
\value{
Summary of the results obtained from esm
}
\description{
The function summary.esm gives a summary from an object from esm()
}
