% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.lmdu.R
\name{summary.lmdu}
\alias{summary.lmdu}
\title{Summarizing Logistic MDU models}
\usage{
\method{summary}{lmdu}(object, ...)
}
\arguments{
\item{object}{An object resulting from lmdu}

\item{\dots}{additional arguments to be passed.}
}
\value{
Summary of the results obtained from lmdu
}
\description{
The function summary.lmdu gives a summary from an object from lmdu()
}
