% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.lpca.R
\name{summary.lpca}
\alias{summary.lpca}
\title{Summarizing Logistic PCA models}
\usage{
\method{summary}{lpca}(object, ...)
}
\arguments{
\item{object}{An object resulting from lpca}

\item{\dots}{additional arguments to be passed.}
}
\value{
Summary of the results obtained from lpca
}
\description{
The function summary.lpca gives a summary from an object from lpca()
}
