% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mcd.R
\name{summary.mcd}
\alias{summary.mcd}
\title{Summarizing an Multinomial Canonical Decomposition Model}
\usage{
\method{summary}{mcd}(object, ...)
}
\arguments{
\item{object}{An object resulting from mcd}

\item{\dots}{additional arguments to be passed.}
}
\value{
Summary of the results obtained from mcd
}
\description{
The function summary.mcd1 gives a summary from an object from mcd()
}
