% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mlr.R
\name{summary.mlr}
\alias{summary.mlr}
\title{Summarizing Multinomial Logistic Regression Model}
\usage{
\method{summary}{mlr}(object, ...)
}
\arguments{
\item{object}{An object resulting from mlr}

\item{\dots}{additional arguments to be passed.}
}
\value{
Summary of the results obtained from mlr
}
\description{
The function summary.mlr gives a summary from an object from mlr()
}
