% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mrrr.R
\name{summary.mrrr}
\alias{summary.mrrr}
\title{Summarizing Multinomial Reduced Rank Model}
\usage{
\method{summary}{mrrr}(object, ...)
}
\arguments{
\item{object}{An object resulting from mrrr}

\item{\dots}{additional arguments to be passed.}
}
\value{
Summary of the results obtained from mrrr
}
\description{
The function summary.mrrr gives a summary from an object from mrrr()
}
