% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mru.R
\name{summary.mru}
\alias{summary.mru}
\title{Summarizing Multinomial Logistic Unfolding model
											   
The function summary.mru gives a summary from an object from mru()}
\usage{
\method{summary}{mru}(object, ...)
}
\arguments{
\item{object}{An object resulting from mru}

\item{\dots}{additional arguments to be passed.}
}
\value{
Summary of the results obtained from mru
}
\description{
Summarizing Multinomial Logistic Unfolding model
											   
The function summary.mru gives a summary from an object from mru()
}
