% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.trioscale.R
\name{summary.trioscale}
\alias{summary.trioscale}
\title{Summarizing TrioScale}
\usage{
\method{summary}{trioscale}(object, ...)
}
\arguments{
\item{object}{An object resulting from trioscale}

\item{\dots}{additional arguments to be passed.}
}
\value{
Summary of the results obtained from trioscale
}
\description{
The function summary.trioscale gives a summary from the MLR in trioscale
}
