% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trioscale.R
\name{trioscale}
\alias{trioscale}
\title{Function for TRIOSCALE}
\usage{
trioscale(y, X)
}
\arguments{
\item{y}{A response formula with 3 classes}

\item{X}{A predictor matrix}
}
\value{
This function returns an object of class trioscale
\item{data}{data}
\item{mlr}{Output object from ts.mlr}
\item{Q}{result from P2Q}
\item{X}{X matrix with coordinates}
\item{Xdf}{X as a data frame}
}
\description{
Function for TRIOSCALE
}
\examples{
\dontrun{
data(diabetes)
output = trioscale(y = diabetes$y, X = diabetes$X)
plot(output)
}


}
