% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twomodedistance.R
\name{twomodedistance}
\alias{twomodedistance}
\title{The function twomodedistance computes the two mode (unfolding) distance}
\usage{
twomodedistance(U, V)
}
\arguments{
\item{U}{An N times S matrix with coordinates in S dimensional Euclidean space.}

\item{V}{An R times S matrix with coordinates in S dimensional Euclidean space.}
}
\value{
D a N by R matrix with Euclidean distances
}
\description{
The function twomodedistance computes the two mode (unfolding) distance
}
